(PC-14598)[PRO] refactor: move adapters into core
