+++ pro/src/core/OfferEducational/adapters/getEducationalStockAdapter.ts
