+++ pro/src/core/OfferEducational/adapters/index.ts
@@ -4,3 +4,6 @@ export * from './getStockCollectiveOfferAdapter'
 export * from './getOfferersAdapter'
 export * from './cancelActiveBookingsAdapter'
 export * from './patchIsOfferActiveAdapter'
+export * from './getEducationalStockAdapter'
+export * from './patchShadowStockAdapter'
+export * from './patchShadowStockIntoEducationalStockAdapter'
