+++ pro/src/core/OfferEducational/utils/extractInitialStockValues.ts
