+++ pro/src/core/OfferEducational/utils/index.ts
@@ -6,3 +6,4 @@ export * from './setInitialFormValues'
 export * from './createStockDataPayload'
 export * from './createPatchStockDataPayload'
 export * from './buildDatetimesForStockPayload'
+export * from './extractInitialStockValues'
