--- pro/src/routes/OfferEducationalStockEdition/types.ts
@@ -1,9 +1,0 @@
-export type StockResponse = {
-  id: string
-  beginningDatetime?: string
-  bookingLimitDatetime?: string
-  price: number
-  numberOfTickets?: number
-  isEducationalStockEditable?: boolean
-  educationalPriceDetail?: string
-}
